Refactor: clean shared library
라이브러리에 필요하지 않은 헤더 파일 정리
