--- Library/include/client_socket.h
@@ -1,2 +1,0 @@
-int create_tcp_sock();
-void connect_with_serv(int client_sock, int port, char *serv_ip_addr);
