--- Library/include/config.h
@@ -1,1 +1,0 @@
-#define BUF_SIZE 256
