--- Library/include/server_socket.h
@@ -1,2 +1,0 @@
-void bind_serv_sock(int serv_sock, int port);
-int accept_and_create_client_sock(int serv_sock);
