+++ tcpsocket_tls/server_socket.c
@@ -1,5 +1,3 @@
-#include "server_socket.h"
-
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
